useUncontrolledInput [nfc]: Move hook to this new file, from ComposeBox
